Update on area change logic for progress bar
Fixes bug that would send user to instruments when clicking on review
