+++ src/interfaces/new-suggestion.ts
@@ -5,3 +5,10 @@ export interface NewSuggestion {
   title: string
   instruments: []
 }
+
+export interface InputsSongInformation {
+  artist: string
+  link: string
+  motivation: string
+  title: string
+}
